Use mixin methods for event context
