+++ server/src/main/java/io/spine/server/route/EventProducers.java
@@ -23,7 +23,6 @@ package io.spine.server.route;
 import com.google.common.collect.ImmutableSet;
 import io.spine.base.EventMessage;
 import io.spine.core.EventContext;
-import io.spine.core.Events;
 
 import java.util.Set;
 
@@ -46,7 +45,7 @@ final class EventProducers {
         @Override
         public Set<I> apply(EventMessage message, EventContext context) {
             @SuppressWarnings("unchecked") // The route creator is responsible for the type check.
-            I id = (I) Events.getProducer(context);
+            I id = (I) context.producer();
             return ImmutableSet.of(id);
         }
