+++ server/src/test/java/io/spine/core/EventsTest.java
@@ -33,15 +33,10 @@ import io.spine.testing.UtilityClassTest;
 import io.spine.testing.client.TestActorRequestFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 import static io.spine.core.Events.checkValid;
-import static io.spine.core.Events.getActor;
-import static io.spine.core.Events.getProducer;
 import static io.spine.core.Events.nothing;
-import static io.spine.protobuf.AnyPacker.unpack;
-import static io.spine.server.type.given.EventsTestEnv.event;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.fail;
@@ -59,7 +54,6 @@ public class EventsTest extends UtilityClassTest<Events> {
             new TestActorRequestFactory(EventsTest.class);
 
     private Event event;
-    private EventContext context;
 
     EventsTest() {
         super(Events.class);
@@ -71,7 +65,6 @@ public class EventsTest extends UtilityClassTest<Events> {
         StringValue producerId = StringValue.of(getClass().getSimpleName());
         EventFactory eventFactory = EventFactory.on(cmd, Identifier.pack(producerId));
         event = eventFactory.createEvent(GivenEvent.message(), null);
-        context = event.context();
     }
 
     private static CommandEnvelope generate() {
@@ -90,27 +83,6 @@ public class EventsTest extends UtilityClassTest<Events> {
               .setDefault(Version.class, Version.getDefaultInstance())
               .setDefault(Event.class, Event.getDefaultInstance())
               .setDefault(ThrowableMessage.class, defaultThrowableMessage);
-    }
-
-    @Nested
-    @DisplayName("given event context, obtain")
-    class GetFromEventContext {
-
-        @Test
-        @DisplayName("actor")
-        void actor() {
-            assertEquals(context.getCommandContext()
-                                .getActorContext()
-                                .getActor(), getActor(context));
-        }
-
-        @Test
-        @DisplayName("producer")
-        void producer() {
-            StringValue msg = unpack(context.getProducerId(), StringValue.class);
-            String id = (String) getProducer(context);
-            assertEquals(msg.getValue(), id);
-        }
     }
 
     @Test
@@ -144,37 +116,5 @@ public class EventsTest extends UtilityClassTest<Events> {
     void acceptGeneratedEventId() {
         EventId eventId = event.getId();
         assertEquals(eventId, checkValid(eventId));
-    }
-
-    //TODO:2019-04-16:alexander.yevsyukov: Move to EventContextTest.
-    // Throw IllegalStateException instead of IllegalArgumentException since it should be from an instance method.
-    @SuppressWarnings({"CheckReturnValue", "ResultOfMethodCallIgnored"})
-    // Method called to throw exception.
-    @Nested
-    @DisplayName("throw IAE when reading tenant ID")
-    class ThrowIaeOnRead {
-
-        @Test
-        @DisplayName("of the event without origin")
-        void forEventWithoutOrigin() {
-            EventContext context = contextWithoutOrigin().build();
-            Event event = event(context);
-            assertThrows(IllegalArgumentException.class, event::tenant);
-        }
-
-        @Test
-        @DisplayName("of the event whose event context has no origin")
-        void forEventContextWithoutOrigin() {
-            EventContext context = contextWithoutOrigin()
-                    .setEventContext(contextWithoutOrigin())
-                    .build();
-            Event event = event(context);
-            assertThrows(IllegalArgumentException.class, event::tenant);
-        }
-
-        private EventContext.Builder contextWithoutOrigin() {
-            return context.toBuilder()
-                          .clearOrigin();
-        }
     }
 }
